Resolved issue with kubeconfig (#126)
* implemented passing the kubeconfig file path via Session

---------

+++ src/executor/job.py
@@ -124,7 +124,7 @@ from abc import ABC, abstractmethod
 from collections import defaultdict
 from itertools import chain
 from pathlib import Path
-from typing import TYPE_CHECKING, Generator, Iterable, cast
+from typing import TYPE_CHECKING, Generator, Iterable, cast, Callable
 
 # import multiprocessing
 import billiard as multiprocessing  # allows to execute child processes from a daemon process
@@ -313,6 +313,10 @@ class PoliciesLoader:
             return GLOBAL_REGION
         return policy.options.region
 
+    @property
+    def _kubeconfig_path(self) -> str | None:
+        return os.getenv(ENV_KUBECONFIG)
+
     def _base_config(self) -> Config:
         match self._cloud:
             case Cloud.AWS:
@@ -342,6 +346,26 @@ class PoliciesLoader:
             log_group='null',
             tracer='default',
         )
+
+    @staticmethod
+    def _kube_session_factory(kubeconfig_path: str) -> Callable:
+        """Returns a factory function that creates Session with config_file"""
+        from c7n_kube.client import Session
+
+        _LOG.debug(f'Creating Kubernetes session factory')
+
+        def factory():
+            return Session(config_file=kubeconfig_path)
+        return factory
+
+    def _session_factory(self) -> Callable | None:
+        """Returns a session factory callable or None"""
+        session_factory = None
+        config_path = self._kubeconfig_path
+        if self._cloud == Cloud.KUBERNETES and config_path:
+            session_factory = self._kube_session_factory(config_path)
+
+        return session_factory
 
     @staticmethod
     def _get_resource_types(policies: list[PolicyDict]) -> set[str]:
@@ -469,6 +493,7 @@ class PoliciesLoader:
         load_resources(self._get_resource_types(policies))
         # here we should probably validate schema, but it's too time-consuming
         provider_policies = defaultdict(list)
+        session_factory = self._session_factory()
         for policy in policies:
             # if policy['name'].endswith(DEPRECATED_RULE_SUFFIX):
             #     _LOG.warning(
@@ -477,7 +502,11 @@ class PoliciesLoader:
             #     )
             #     continue
             try:
-                pol = Policy(policy, options)
+                pol = Policy(
+                    data=policy,
+                    options=options,
+                    session_factory=session_factory,
+                )
             except PolicyValidationError:
                 _LOG.warning(
                     f'Cannot load policy {policy["name"]} '
@@ -1241,11 +1270,11 @@ def get_platform_credentials(job: Job, platform: Platform) -> dict | None:
         config.add_user(user, token)
         config.add_context(context, cluster, user)
         config.current_context = context
-        return {ENV_KUBECONFIG: config.raw}
+        return {ENV_KUBECONFIG: str(config.to_temp_file())}
     elif kubeconfig:
         _LOG.debug('Only kubeconfig is provided')
         config = Kubeconfig(kubeconfig)
-        return {ENV_KUBECONFIG: config.raw}
+        return {ENV_KUBECONFIG: str(config.to_temp_file())}
     if platform.type != PlatformType.EKS:
         _LOG.warning('No kubeconfig provided and platform is not EKS')
         return
@@ -1301,7 +1330,7 @@ def get_platform_credentials(job: Job, platform: Platform) -> dict | None:
         ca=cluster['certificateAuthority']['data'],
         token=TokenGenerator(sts).get_token(platform.name),
     )
-    return {ENV_KUBECONFIG: token_config.build_config()}
+    return {ENV_KUBECONFIG: str(token_config.to_temp_file())}
 
 
 def get_rules_to_exclude(tenant: Tenant) -> set[str]:
@@ -1457,23 +1486,12 @@ def multi_account_event_driven_job() -> int:
 
 def job_initializer(
     envs: dict,
-    cloud: Cloud,
 ):
     _LOG.info(
         f'Initializing subprocess for a region: {multiprocessing.current_process()}'
     )
-    try:
-        if cloud is Cloud.KUBERNETES and ENV_KUBECONFIG in envs:
-            envs = BSP.credentials_service.k8s_credentials_to_file(
-                envs[ENV_KUBECONFIG]
-            )
-
-        envs = {str(k): str(v) for k, v in envs.items() if v}
-
-        os.environ.update(envs)
-        os.environ.setdefault('AWS_DEFAULT_REGION', AWS_DEFAULT_REGION)
-    except Exception as e:
-        _LOG.exception(f'Unexpected error occurred during job initialization {e}')
+    os.environ.update(envs)
+    os.environ.setdefault('AWS_DEFAULT_REGION', AWS_DEFAULT_REGION)
 
 
 def process_job_concurrent(
@@ -1532,6 +1550,12 @@ def process_job_concurrent(
         filename := os.environ.get(ENV_AZURE_CLIENT_CERTIFICATE_PATH)
     ):
         _LOG.debug(f'Removing temporary azure certificate file {filename}')
+        Path(filename).unlink(missing_ok=True)
+
+    if cloud is Cloud.KUBERNETES and (
+        filename := os.environ.get(ENV_KUBECONFIG)
+    ):
+        _LOG.debug(f'Removing temporary kubeconfig file {filename}')
         Path(filename).unlink(missing_ok=True)
 
     return runner.n_successful, runner.failed
@@ -1803,6 +1827,8 @@ def run_standard_job(ctx: JobExecutionContext):
     if credentials is None:
         raise ExecutorException(ExecutorError.NO_CREDENTIALS)
 
+    credentials = {str(k): str(v) for k, v in credentials.items() if v}
+
     policies = list(
         skip_duplicated_policies(
             ctx=ctx,
@@ -1828,7 +1854,7 @@ def run_standard_job(ctx: JobExecutionContext):
         with multiprocessing.Pool(
             processes=1,
             initializer=job_initializer,
-            initargs=(credentials, cloud,),
+            initargs=(credentials,),
         ) as pool:
             pair = pool.apply(
                 process_job_concurrent, (policies, ctx.work_dir, cloud, region)
